# Ques_backend/docs/api/api_specification.yaml
openapi: 3.0.3
info:
  title: Ques Backend API
  description: |
    Complete API specification for the Ques application backend.
    Includes AI-powered user discovery, chat system, swipe mechanics, payments, and more.
  version: 1.0.0
  contact:
    name: Ques Backend Team
    email: backend@ques.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ques.app
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /whispers:
    get:
      tags:
        - Whispers
      summary: Get received whispers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, declined, expired]
      responses:
        '200':
          description: Whispers retrieved
...
  /whispers/settings:
    get:
      tags:
        - Whispers
      summary: Get whisper settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Settings retrieved

    put:
      tags:
        - Whispers
      summary: Update whisper settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                wechatId:
                  type: string
                customMessage:
                  type: string
                autoAccept:
                  type: boolean
      responses:
        '200':
          description: Settings updated

  # Payment endpoints
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Notifications retrieved

  # Contact endpoints
...
  /user-settings:
    get:
      tags:
        - User Settings
      summary: Get user settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Settings retrieved

    put:
      tags:
        - User Settings
      summary: Update user settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notifications:
                  type: object
                preferences:
                  type: object
      responses:
        '200':
          description: Settings updated

  # AI Services endpoints
...